Add stash_package_dir, giving option to skip package download
Servers may not have access to download files from the Internet,
so this option (replacing internal tmpdir) allows a directory
to be specified that will be checked for the installation packages.

If they are found they will be used, otherwise download will be
attempted using the existing wget method, saving files to this
directory.

Defaults to /var/tmp instead of /tmp (previous value of tmpdir).